Add optional pattern to filter file types. (#4)
I really like this action, but I've found that I often want to filter the outputs that it produces and hand them to other tasks such as linters. This PR adds an optional pattern input which is a regex that defaults to .* that can be used to filter the files that end up in each array.