Add optional `pr-number` input (#8)
This optional input allows users to run this action outside of the pull_request event context.

Certain pipelines are not conducive to running solely inside the pull_request event. For example, we are creating a custom deployment event triggered by a Netlify webhook when the build succeeds on their end. After this event triggers, we'd then like to use the changed files of the PR alongside the deployed preview url.