also output the count of files changed/deleted/created (#29)
I had a setup where I wanted to switch the behavior of subsequent steps depending on some threshold of changed files. Counting those from the JSON/plain output failed because the output was too big to display/run in the action step. So I added the counting+output directly in the action.
If you want to include this, I'd be happy to change naming, etc.